refactor(firestore): internalize value codec
Replace the accidental pyric/firestore-values product surface with the explicit pyric/firestore/internal/value-codec worker seam. Preserve the leaf codec behavior and update package consumers, architecture docs, and API-reference inventory.
