+++ packages/site-docs/test/api-reference.test.ts
@@ -47,9 +47,9 @@ describe('generated API reference inventory', () => {
   });
 
   test('uses unique stable routes backed by declaration entries', () => {
-    expect(descriptors).toHaveLength(49);
-    expect(new Set(descriptors.map(({ importPath }) => importPath)).size).toBe(49);
-    expect(new Set(descriptors.map(({ slug }) => slug)).size).toBe(49);
+    expect(descriptors).toHaveLength(48);
+    expect(new Set(descriptors.map(({ importPath }) => importPath)).size).toBe(48);
+    expect(new Set(descriptors.map(({ slug }) => slug)).size).toBe(48);
     for (const descriptor of descriptors) {
       expect(existsSync(descriptor.typesPath), descriptor.importPath).toBeTrue();
       expect(descriptor.slug).toMatch(/^[a-z0-9-]+-reference-api$/);
